# api-spec/src/main/resources/static/api.yml
...
swagger: '2.0'
info:
  title: Name of the App API
  description: Use our services using this API
  version: "1.0.0"
# the domain of the service
host: api.app.com
...
schemes:
  - https
# will be prefixed to all paths
basePath: /v1
produces:
  - application/json
paths:
...
  /notifications:
    post:
      summary: Create notification
      description: |
        This method creates a new notification.
      parameters:
        - name: notification_id
          in: query
          description: Unique identifier representing a specific notification.
          required: true
          type: string
        - name: creator
          in: query
          description: Creator of the notification.
          required: true
          type: string
        - name: title
          in: query
          description: Title of the notification.
          required: true
          type: string
        - name: body
          in: query
          description: Body of the notification.
          required: true
          type: string
      tags:
        - Notifications
      responses:
        201:
          description: Notification created
          schema:
            $ref: '#/definitions/Notification'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    get:
      summary: Get an array of notifications
      description: |
        This method returns an array of those notifications that haven't been received yet by a certain user.
      parameters:
        - name: user
          in: query
          description: Owner of the notifications to be returned.
          required: true
          type: string
      tags:
        - Notifications
      responses:
        200:
          description: Notifications that haven't been received yet
          schema:
            type: array
            items:
              $ref: '#/definitions/Notification'
        404:
          description: Notification not found
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /notifications/{notification_id}:
    put:
      summary: Modify notification
      description: |
        This method modifies the data of a previously created notification.
      parameters:
        - name: notification_id
          in: path
          description: Unique identifier representing a specific notification.
          required: true
          type: string
        - name: creator
          in: query
          description: Creator of the notification.
          required: true
          type: string
        - name: title
          in: query
          description: Title of the notification.
          required: false
          type: string
        - name: body
          in: query
          description: Body of the notification.
          required: false
          type: string
      tags:
        - Notifications
      responses:
        202:
          description: Notification's data modified
          schema:
            $ref: '#/definitions/Notification'
        304:
          description: Notification's data not modified
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Notification not found
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      summary: Delete notification
      description: |
        This method deletes the data of a previously created notification.
      parameters:
        - name: notification_id
          in: path
          description: Unique identifier representing a specific notification.
          required: true
          type: string
        - name: creator
          in: query
          description: Creator of the notification.
          required: true
          type: string
      tags:
        - Notifications
      responses:
        200:
          description: Notification deleted
          schema:
            $ref: '#/definitions/Notification'
        404:
          description: Notification not found
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    get:
      summary: Get notification by notification_id
      description: |
        This method returns a notification by notification_id.
      parameters:
        - name: notification_id
          in: path
          description: Unique identifier representing a specific notification.
          required: true
          type: string
        - name: creator
          in: query
          description: Creator of the notification.
          required: true
          type: string
      tags:
        - Notifications
      responses:
        200:
          description: Notification with notification_id requested
          schema:
            $ref: '#/definitions/Notification'
        404:
          description: Notification not found
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  Notification:
    type: object
    properties:
      notification_id:
        type: string
        description: Unique identifier representing a specific notification.
      creator:
        type: string
        description: Creator of the notification.
      title:
        type: string
        description: Title of the notification.
      body:
        type: string
        description: Body of the notification.
...
  Error:
    type: object
    properties:
      title:
        type: string
      body:
        type: string